requirements.txt
google-generativeai==0.5.2
python-dotenv==1.0.1
google-search-results==2.4.2
requests==2.31.0


Flask==2.3.2
reportlab==4.1.0
python-dotenv==1.0.1
google-generativeai==0.5.2

